Fixed models. Added participant roles seeder. Automatically adds creator to creator of project.

+++ database/migrations/2024_06_27_094217_add_profile_picture_to_users_table.php
@@ -22,7 +22,10 @@ return new class extends Migration
     public function down(): void
     {
         Schema::table('users', function (Blueprint $table) {
-            $table->dropColumn('profile_picture');
+            // Check if the column exists before attempting to drop it
+            if (Schema::hasColumn('users', 'profile_picture')) {
+                $table->dropColumn('profile_picture');
+            }
         });
     }
 };
